--- src/modules/crypto/index.d.ts
@@ -1,6 +1,0 @@
-export type Encrypted = string;
-
-export type Decrypted = {
-    message: string;
-    signature: string;
-}
