+++ src/routes/account/Register.svelte
@@ -8,20 +8,14 @@
 	import Info from '@@Components/Info.svelte';
 	import { users } from '@@Stores/users';
 	import { request } from '@@Modules/nerves';
+	import { generateKeyPair } from '@@Modules/crypto';
 
-  import type { InfoType } from '@@Components/Info';
-  import type { Keypair } from '@@Modules/user';
+    import type { InfoType } from '@@Components/Info';
 
 	let info: InfoType;
     $: info = { value: '', type: '', loading: false };
 
 	const rsa = new RSA();
-	
-	const generateKeyPair = (): Promise<Keypair> => {
-    return new Promise((resolve) => {
-      rsa.generateKeyPair(resolve);
-		});
-	};
 
 	const register = async (e: Event) => {
 		e.preventDefault();
